+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - PHPUnit attributes `Group` and `DataProvider` are now longer read when running tests with BlackBox, use the `Innmind\BlackBox\` prefixed ones instead
 - All proofs will fail if they throw an exception (instead of stopping the script)
 - `Innmind\BlackBox\Runner\Printer` is now a final class
+- `Innmind\BlackBox\Set\MadeOf` have been moved to `Innmind\BlackBox\Set\Provider\Strings\MadeOf`
 
 ### Removed